rewrite to support polygons
This is a significant set of changes to support joints on
non-90-degree edges, including on polygons having other than 4 sides;
some examples are in polygons-test.scad.  These changes require a
slightly different calling convention for all joint types when
describing joints, and therefore break backward compatibility.  We
haven't yet decided whether this should be a new major version or a
completely different package name.

The new calling convention's main feature is that we pass the edge
number (numbered from zero, as listed in the points[] array), instead
of passing in the edge angle -- this deprecates the UP, DOWN, LEFT,
and RIGHT globals.

As of this commit, there is still some work that needs to be done to
fix lasercutoutBox, but boxes can still be constructed "by hand" using
lasercutoutSquare or lasercutout.

The README still needs to be updated as well.